+++ deno.json
@@ -4,21 +4,35 @@
     "allowJs": true,
     "lib": ["deno.window"]
   },
+  "unstableSloppy": true,
   "imports": {
     "@/": "./src/",
+    "@/types/": "./src/types/",
+    "@/src/": "./src/",
+
+    "@/src/index": "./src/index.ts",
+    "@/src/client": "./src/client.ts",
+
     "@/types/types": "./src/types/types.ts",
+    "@/types/index": "./src/types/index.ts",
+
     "@/types/discipline/meteorologicalProducts/categories": "./src/types/discipline/meteorologicalProducts/categories.ts",
     "@/types/discipline/meteorologicalProducts/types": "./src/types/discipline/meteorologicalProducts/types.ts",
     "@/types/discipline/meteorologicalProducts/momentum": "./src/types/discipline/meteorologicalProducts/momentum.ts",
+    "@/types/discipline/meteorologicalProducts/index": "./src/types/discipline/meteorologicalProducts/index.ts",
+
     "@/types/discipline/oceanographicProducts/categories": "./src/types/discipline/oceanographicProducts/categories.ts",
     "@/types/discipline/oceanographicProducts/waves": "./src/types/discipline/oceanographicProducts/waves.ts",
     "@/types/discipline/oceanographicProducts/currents": "./src/types/discipline/oceanographicProducts/currents.ts",
-    "@/types/discipline/oceanographicProducts/types": "./src/types/discipline/oceanographicProducts/types.ts",
     "@/types/discipline/oceanographicProducts/ice": "./src/types/discipline/oceanographicProducts/ice.ts",
+    "@/types/discipline/oceanographicProducts/types": "./src/types/discipline/oceanographicProducts/types.ts",
+    "@/types/discipline/oceanographicProducts/index": "./src/types/discipline/oceanographicProducts/index.ts",
+
     "@/types/discipline/units/types": "./src/types/discipline/units/types.ts",
     "child_process": "node:child_process",
     "util": "node:util",
     "readline": "node:readline",
+
     "vitest": "npm:vitest"
   }
 }
