+++ nbs/241215_queries.nb.ts
@@ -1,6 +1,6 @@
 
 //#nbts@code
-import { EccodesWrapper } from "../src/index.ts";
+import { EccodesWrapper } from "../src/client.ts";
 
 //#nbts@code
 const datapath = `./data/gefs.wave.grib2`;
