+++ src/types/discipline/meteorologicalProducts/index.ts
@@ -1,0 +1,3 @@
+export * from "@/types/discipline/meteorologicalProducts/categories";
+export * from "@/types/discipline/meteorologicalProducts/momentum";
+export * from "@/types/discipline/meteorologicalProducts/types";
