+++ src/types/discipline/oceanographicProducts/index.ts
@@ -1,0 +1,5 @@
+export * from "@/types/discipline/oceanographicProducts/categories";
+export * from "@/types/discipline/oceanographicProducts/waves";
+export * from "@/types/discipline/oceanographicProducts/currents";
+export * from "@/types/discipline/oceanographicProducts/ice";
+export * from "@/types/discipline/oceanographicProducts/types";
