+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "@/types/discipline/meteorologicalProducts";
+export * from "@/types/discipline/oceanographicProducts";
+export * from "@/types/discipline/units/types";
